+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Changed
+
+- Don't add temp prefix when using `extract` command.
+- VS Code extension: mirrord enable/disable to be per workspace.
+- VS Code extension: bundle the resources
+
 ## 3.13.4
 
 ### Changed